requirements.txt
click==8.0.1
flake8==3.9.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
SQLAlchemy==1.4.22
Werkzeug==2.0.1
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.1
flake8==3.9.2
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
Werkzeug==2.0.1
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.1
flake8==3.9.2
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
Werkzeug==2.0.1
